Main: Add commands when running the program
I just prefer this to flags. Added a command for 'launchd' to install a
Launch Agent, 'start' to start running the program', and 'help' which
prints a usage screen.

Since starting the tracker requires passing in the 'start' command, the
Launch Agent plist has been updated to accept the 'start' command.

Additionally, since the command to start the tracker is 'start', the
function `runTracker` has been renamed to `startTracker`.