Main: Pass the address of the tracker pointer since it can change
There was a bug where a new tracker wouldn't be created when the day
changed. What was needed was to pass a pointer to the Tracker pointer
so that we can change the underlying tracker.

Similarly, since all of the other functions called in `startTracker`
originally accepted a tracker pointer as a parameter, we needed to
update those functions to accept a **tr.Tracker too so that the
underlying tracker reference could change.